+++ plugins/06_WarlordsOfDraenor/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 * Initial release!
 
+
+* Draenor
+  * [Voidtalon](https://www.wowhead.com/item=121815/voidtalon-of-the-dark-star)
+
+
+* Gorgrond
+  * [Poundfist](https://www.wowhead.com/npc=50985/poundfist)
+  * Rares for [Gorgrond Monster Hunter](https://www.wowhead.com/achievement=9400/gorgrond-monster-hunter) achievement
+
+
 * Tanaan Jungle
   * Rares for [Jungle Stalker](https://www.wowhead.com/achievement=10070/jungle-stalker) achievement
   * Pet Battles for [Tiny Terrors in Tanaan](https://www.wowhead.com/achievement=10052/tiny-terrors-in-tanaan) achievement